+++ src/plugins/memory/memory-types.ts
@@ -1,4 +1,9 @@
-import type { PreparedQuery, RxDocumentData, RxStorage } from '../../types';
+import type {
+    PreparedQuery,
+    RxAttachmentWriteData,
+    RxDocumentData,
+    RxStorage
+} from '../../types';
 
 export type RxStorageMemorySettings = {};
 export type RxStorageMemoryInstanceCreationOptions = {};
@@ -28,6 +33,11 @@ export type MemoryStorageInternals<RxDocType> = {
      */
     removed: boolean;
     documents: Map<string, RxDocumentData<RxDocType>>;
+    /**
+     * Attachments data, indexed by a combined string
+     * consisting of [documentId + '||' + attachmentId]
+     */
+    attachments: Map<string, RxAttachmentWriteData>
     byIndex: {
         /**
          * Because RxDB requires a deterministic sorting